Moved half of SP to keep in thumbs repo. Other half is getting moved to core.
